progress_bar: updated the color scheme, but i have to admit it does looks kinda boring.
spinner: re-did the spinner widget for various reasons. 
I think it was too cramped, trying to do too much for its $2.00 
dollars. 

I still don't like its 'image' implementation. We shouldn't 
have to draw the spinner and then still have the buttons above 
it. I believe the spinner and the entry box should be drawn 
separate, thus allowing more flexibility. 

Simon, you said the entry box in the spinner should be the same 
size as the entry widget. I know you are working on the entry 
widget so that the size gets calculated by the library and 
not defined by the theme. Another reason, why I re-did the 
widget. I think no matter the outcome the up/down buttons in 
the spinner will still be too small.

I still plan to redo the buttons on the spinner, I borrowed 
these from an elicit theme, which uses an outdated gradient 
color scheme. Perhaps redo the spinner completely so the 
buttons are not part of 'spinner.png'. And again, Simon, if you 
don't like the spinner let me know ASAP and we'll do something 
else.