* Fix the dialog separator * Make the background of the tree a bit darker. It looks better when the tree is almost empty imho (the contrast is lower) * Lower the entry's text by 1px * Remove text-classes for now since they makes edje-object loading really slow (Bug in Edje probably)
Note for Alberto: please don't revert your spinner changes, I start to 
get used to it and I have to say I really like it now :) Only thing is, 
the background of the spinner is darker than the background of the 
entry, and it may look bad if those two widgets are close. See this 
screenshot: http://mtreny.free.fr/detour_entry_spinner.png . Don't know 
which of the entry or of the spinner should be changed.