Added status bar widget. Very simple, could be improved. Any ideas?
Simon: Tree widget row was overlapping the tree_over image, so 
i had to move it -1 to the left (see Line: 216 tree.edc), this 
leaves as 1px space between the scrollbar and the tree, 
I don't know if this is acceptable but whatever (only visible 
when the scrollbar kicks in). The other solution is to redraw 
the over image.