Here is what I added:
When you click either the screenshot or the color palette 
window a screen wipe effect appears. For this to work
you right click on either window. Or right click on the 
window icon (color picker, etc)

Right clicking on the screenshot window will send the 
screenshot to the gimp via gimp-remote or open the gimp
for you if you don't currently have the gimp open.

I have a 7 button mouse, so the map pointers may not be the 
same for everybody. For example, "mouse,clicked,3" is the 
right button on my mouse. 

If you click on the hex trippled (#123456) it will copy the 
value into the clipboard. 

Right click on color paletes to delete them.

Enjoy!