Animated progress bar.
Simon: a few months ago Codewarrior and I tried to created an animated 
progress bar, and the results varied a lot. The animation was always 
jerky and we couldn't figure it out (well maybe I should speak for 
myself). Anyways here is an another attempt, it is much better than the 
previous attemps, but it still is not in sync. 

What could it be? missing a fill ? Too many images ? (yeah 32 total) 
will this consume too much CPU ?

Even if we manage to fix it and decide that we don't want to use it, 
maybe we could leave in as an #define if/else logic for those that want 
the animation.