Disabled focus hints momentarily, working out on a new design.
This is a quick commit.

Updating buttons to reflect detour-e17's latest changes.

Radio/Check buttons disabled state needs to be updated and check 
that disabled text actually matches.... nvm

Simon: i copied the toggle code over to the combobox and updated 
it to reflect the combo box, yet the toggle doesn't seem to work 
properly. I went over it for a while, and the problem seems to 
start on Line:190 -- it beats me.