Edje group aliases are great! Fixing the last of the formating changes. Add support to e/fileman/scrollframe/desktop
Exebuf: anyone wanting to re-design exebuf ?

Simon: i'm glad you spoted the ibar icon fade, the reason why 
they are fade is so that the overlay Icon Label text description 
is easier to read, but there is no reason for them to be faded if 
the user had the Icon Labels disabled. Should we just set the 
opacity to 255 255 255 255 ? Its up for grabs.