Updates to the tree images.
Note: the tree header label (text) needs to be centered 
properly. The tree headers expand based on the location of the 
header label (text). I didn't have time to go back and edit the 
tree_header macros, not sure if setting the size of the headers 
to fixed is such a good idea.

Slider, Radio and Check widget updates (matching detour-e17's 
latest updates) I didn't have the patience to format the code, 
sorry about that.

The scrollbars have been modified to fit with the tree widget.

Simon: radio and check widgets now have a different style of 
focus, their labels no longer need to change color. I forgot 
where to edit this.

Another note, the scrollbar have a discrepency when both are 
visible. Please see the following url: 
http://www.cored.org/tmp/scrollb_vis.png