Check, Radio, and Seeker widget updates: - new images - new focus behaviour (see detour-e17)
Jaime: see http://www.cored.org/tmp/new_focus_behaviour.png

Old focus behaviour still there. This is also true for the 
seeker.

Note to self: the seeker should be confined to the rail, not to 
the beginning and end of the image(s).

Tree header image updates. On etk, the tree_headers are outside 
the tree, while in ewl they are inside ?