Listen up.
This is a "WORK IN PROGRESS". In order to make it easier
for others to read the source code, I will be centralizing
every widget to 1 file. So instead of having:

example.edc \
example_main.edc \
example_programs.edc

We'll have:
example.edc

Started documenting a couple of files. So much more to do.

Improved cosmetics.